addding authorization checks

+++ controllers/feed.js
@@ -149,6 +149,12 @@ exports.updatePost = (req, res, next) => {
         throw error;
       }
 
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authenticated');
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -183,7 +189,11 @@ exports.deletePost = (req, res, next) => {
         throw error;
       }
 
-      // Checked logged in
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error('Not authenticated');
+        error.statusCode = 403;
+        throw error;
+      }
 
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
